+++ CHANGELOG.md
@@ -1,4 +1,4 @@
 ### Changes
-- Made behavior of uninsulated top ovens more consistent. Now they will heat items, but only to half the requested temperature.
-- Added a note about the max greenhouse size to the book.
-- Replaced uses of the Empty Jar With Lid item with a tag internally.
+- Added Stainless Steel jar lids
+- Updated the stomping barrel recipe again.
+- Made the check for flowers for bees much more forgiving, most plants should work now.